refactor(datepicker): reutiliza po-calendar publico
Fixes DTHFUI-4615
