--- projects/ui/src/lib/components/po-field/po-datepicker/po-calendar/po-calendar.component.html
@@ -1,88 +1,0 @@
-<div class="po-calendar-overlay" [class.po-invisible]="overlayInvisible"></div>
-<div *ngIf="visible" [ngClass]="setMobileVisualization()">
-  <div *ngIf="dayVisible" #days>
-    <div class="po-calendar-header">
-      <span (click)="onPrevMonth()" class="po-calendar-header-left po-icon po-icon-arrow-left"></span>
-      <div class="po-calendar-header-title" attr-calendar>
-        <span class="po-clickable po-mr-1" (click)="selectMonth()">{{ displayMonth }}</span>
-        <span class="po-clickable" (click)="selectYear()">{{ displayYear }}</span>
-      </div>
-      <span (click)="onNextMonth()" class="po-calendar-header-right po-icon po-icon-arrow-right"></span>
-    </div>
-    <div class="po-calendar-content">
-      <div class="po-calendar-labels">
-        <div class="po-calendar-label" *ngFor="let day of displayWeedDays">
-          {{ day.toLowerCase() }}
-        </div>
-      </div>
-      <div class="po-calendar-content-list-day">
-        <div
-          *ngFor="let day of displayDays"
-          class="po-calendar-day po-calendar-box-background"
-          (click)="onSelectDate(day)"
-          [ngClass]="getDayBackgroundColor(day)"
-        >
-          <span *ngIf="day !== 0" [ngClass]="getDayForegroundColor(day)">
-            {{ day.getDate() }}
-          </span>
-        </div>
-      </div>
-    </div>
-  </div>
-  <div *ngIf="monthVisible" #months>
-    <div class="po-calendar-header">
-      <span (click)="updateYear(-1)" class="po-calendar-header-left po-icon po-icon-arrow-left"></span>
-      <div class="po-calendar-header-title" attr-calendar>
-        <span class="po-clickable" (click)="selectYear()">{{ displayYear }}</span>
-      </div>
-      <span (click)="updateYear(1)" class="po-calendar-header-right po-icon po-icon-arrow-right"></span>
-    </div>
-    <div class="po-calendar-content">
-      <div class="po-calendar-labels">
-        <div class="po-calendar-label">
-          {{ getWordMonth() }}
-        </div>
-      </div>
-      <div class="po-calendar-content-list-month">
-        <div
-          *ngFor="let month of displayMonths; let i = index"
-          class="po-calendar-month"
-          (click)="onSelectMonth(displayYear, i)"
-          [ngClass]="getBackgroundColor(i, displayMonthNumber)"
-          attr-calendar
-        >
-          <span [ngClass]="getForegroundColor(i, displayMonthNumber)">
-            {{ month }}
-          </span>
-        </div>
-      </div>
-    </div>
-  </div>
-  <div *ngIf="yearVisible" #years>
-    <div class="po-calendar-header">
-      <span (click)="updateYear(-10)" class="po-calendar-header-left po-icon po-icon-arrow-left"></span>
-      <div class="po-calendar-header-title">{{ displayStartDecade }} - {{ displayFinalDecade }}</div>
-      <span (click)="updateYear(10)" class="po-calendar-header-right po-icon po-icon-arrow-right"></span>
-    </div>
-    <div class="po-calendar-content">
-      <div class="po-calendar-labels">
-        <div class="po-calendar-label">
-          {{ getWordYear() }}
-        </div>
-      </div>
-      <div class="po-calendar-content-list-year">
-        <div
-          *ngFor="let year of displayDecade; let i = index"
-          class="po-calendar-year"
-          (click)="onSelectYear(year, displayMonthNumber)"
-          [ngClass]="getBackgroundColor(year, currentYear)"
-          attr-calendar
-        >
-          <span [ngClass]="getForegroundColor(year, currentYear)">
-            {{ year }}
-          </span>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
