+++ projects/ui/src/lib/components/po-field/po-datepicker/po-datepicker.module.ts
@@ -2,10 +2,10 @@ import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { NgModule } from '@angular/core';
 
+import { PoCalendarModule } from '../../po-calendar/po-calendar.module';
 import { PoFieldContainerModule } from '../po-field-container/po-field-container.module';
 import { PoCleanModule } from '../po-clean/po-clean.module';
 
-import { PoCalendarComponent } from './po-calendar/po-calendar.component';
 import { PoDatepickerComponent } from './po-datepicker.component';
 
 /**
@@ -14,8 +14,8 @@ import { PoDatepickerComponent } from './po-datepicker.component';
  * Módulo do componente `po-datepicker`.
  */
 @NgModule({
-  imports: [CommonModule, FormsModule, PoFieldContainerModule, PoCleanModule],
+  imports: [CommonModule, FormsModule, PoFieldContainerModule, PoCleanModule, PoCalendarModule],
   exports: [PoDatepickerComponent],
-  declarations: [PoDatepickerComponent, PoCalendarComponent]
+  declarations: [PoDatepickerComponent]
 })
 export class PoDatepickerModule {}
